Big update
Fixed an issue with apply_mask, and changed it from throwing to returning a Boolean.

Changed propagate_solution from throwing to returning a Boolean.

Added comparison operators to sudoku_board.

Added a read function that is independent of the agenda used to later solve it.

Added a make_mask function.

Axed the old solvers.